Detect status bar carrier state when theme change.
This will fix the issue when changing or applying theme,
the status bar carrier label was set to enabled.
Then when we want to disable the status bar carrier label
thru Settings the systemui would crash.

now we let the system detect if the user has
enabled or disabled the carrier label when we change or apply theme.

Change-Id: I062ee85f3b4912ebecfb5e1e7ad6e4d222f044fa

Conflicts:

	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java